Structure test components better
